internal: Move IsAllowedGitCmd from vcs/git to gitserver/gitdomain
